feat: NotificationDTO
