+++ src/main/java/site/matzip/notification/controller/NotificationController.java
@@ -8,7 +8,9 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import site.matzip.config.auth.PrincipalDetails;
-import site.matzip.notification.entity.Notification;
+import site.matzip.member.domain.Member;
+import site.matzip.member.service.MemberService;
+import site.matzip.notification.dto.NotificationDTO;
 import site.matzip.notification.service.NotificationService;
 
 import java.util.List;
@@ -17,15 +19,17 @@ import java.util.List;
 @RequestMapping("/usr/notification")
 @RequiredArgsConstructor
 public class NotificationController {
+    private final MemberService memberService;
     private final NotificationService notificationService;
 
     @GetMapping("/list")
     @PreAuthorize("isAuthenticated()")
     public String showList(Model model, @AuthenticationPrincipal PrincipalDetails principalDetails) {
+        Member member = memberService.findByUsername("user1").orElseThrow();
 
-        List<Notification> notifications = notificationService.getNotifications(principalDetails.getMember());
+        List<NotificationDTO> notificationDTOS = notificationService.convertToNotificationDTOS(member);
 
-        model.addAttribute("notifications", notifications);
+        model.addAttribute("notificationDTOS", notificationDTOS);
 
         return "usr/notification/list";
     }
