+++ src/main/java/site/matzip/notification/entity/Notification.java
@@ -9,7 +9,6 @@ import site.matzip.base.domain.BaseEntity;
 import site.matzip.member.domain.Member;
 
 import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
@@ -37,19 +36,5 @@ public class Notification extends BaseEntity {
         notification.fromMember = fromMember;
 
         return notification;
-    }
-
-    public String getAfterCreateNotification() {
-        long diff = ChronoUnit.SECONDS.between(getCreateDate(), LocalDateTime.now());
-        if (diff < 60) return diff + "초";
-        else if (diff < 3600) {
-            return (diff / 60) + "분";
-        } else if (diff < 86400) {
-            return (diff / 60 / 60) + "시간";
-        } else return (diff / 60 / 60 / 24) + "일";
-    }
-
-    public void setAfterReadNotification(LocalDateTime localDateTime) {
-        this.readDate = localDateTime;
     }
 }
